refactor: collapse single-purpose modules into the files that use them
Four small structural cleanups, no behavior change. cargo build
and cargo clippy --release pass clean.

1. src/keys.rs was misnamed after the LUKS removal — the module
   was one function (wait_for_config) with a single caller in
   main.rs. Inline into main.rs alongside the matching subcommand
   definition; delete keys.rs. Renamed the function to
   wait_for_and_persist_config since it both runs the HTTP wait
   and writes the result to disk; "wait_for_config" alone
   undersold the persist step.

2. src/persistence.rs held one constant (PERSISTENT_CONFIG_FILE)
   and one function (write_persistent_config) used only by the
   above. Inline into main.rs; the orchestrator and the IO it
   does live next to each other now.

3. src/utils/file.rs held three helpers (write_file_with_perms,
   set_file_permissions, create_dir_safe) used only by the now-
   inlined persistence code. Inline at the call site;
   set_file_permissions only existed because write_file_with_perms
   wanted to compose, both collapse into the four-line
   write/set-perms sequence write_persistent_config actually
   needed. The utils/ directory disappears.

4. src/server/{mod,http,handlers,validation}.rs was four files
   for one logical unit (axum HTTP listener + handler + payload
   validator), ~150 LoC total. Combined into a single
   src/server.rs. AppState and the validate_* helpers were `pub`
   only to cross the old submodule boundary; they become private
   file-local now. The validation logic itself is unchanged — its
   retirement (along with the args/log fields it gates) is the
   next commit.

Also added a TODO above PERSISTENT_CONFIG_MODE: ownership and
permissions on /persistent/conf/node.json should be a seismic-images
concern (User= + ExecStartPre in tdx-init.service), not baked into
the binary. Today tdx-init runs as root and we chmod to 0o644
explicitly; the cleaner shape is to run tdx-init as a tdx-init:eth
system user and let the umask produce 0o644 naturally.

src/ goes from 11 files across 3 directories to 4 flat files.
Net -57 LoC.
