--- src/server/mod.rs
@@ -1,3 +1,0 @@
-pub mod handlers;
-pub mod http;
-pub mod validation;
